EKDOM CHUP

+++ queries/deleteQueries.js
@@ -1,18 +1,32 @@
 const query = require("./queries.js");
 
-async function deleteUserCascade(tableName, id) {
-	const q = `DELETE FROM ${tableName} WHERE ID=:1`;
+async function deleteTableRow(tableName, id) {
+	let q = "";
+	if (tableName != "people" || tableName != "PEOPLE") {
+		q = `DELETE FROM ${tableName} WHERE ID=:1`;
+	} else {
+		q = `BEGIN
+          PEOPLE_DELETE(:1);
+        END;`;
+	}
 	const params = [id];
 
 	return await query.db_query(q, params);
 }
 
-async function deleteUserCascadeCustom(tableName, attribute, attrValue) {
-	const q = `DELETE FROM ${tableName} WHERE ${attribute}=:1`;
+async function deleteTableRowCustom(tableName, attribute, attrValue) {
+	let q = "";
+	if (tableName != "people" || tableName != "PEOPLE") {
+		q = `DELETE FROM ${tableName} WHERE ${attribute}=:1`;
+	} else {
+		q = `BEGIN
+          PEOPLE_DELETE(:1);
+        END;`;
+	}
 	const params = [attrValue];
 
 	return await query.db_query(q, params);
 }
 
-exports.deleteUserCascade = deleteUserCascade;
-exports.deleteUserCascadeCustom = deleteUserCascadeCustom;
+exports.deleteTableRow = deleteTableRow;
+exports.deleteTableRowCustom = deleteTableRowCustom;
